Split pretty functions
