+++ lib/theory/src/Pretty.hs
@@ -1,0 +1,53 @@
+module Pretty (
+    module Pretty
+) where
+
+import           Prelude                             hiding (id, (.))
+
+
+-- import           Data.Typeable
+import           Data.List
+
+import           Control.Category
+
+-- import qualified Data.Label.Total
+
+
+import           Theory.Model
+import           Theory.Proof
+import           Theory.Text.Pretty
+
+
+------------------------------------------------------------------------------
+-- Pretty printing
+------------------------------------------------------------------------------
+
+-- | Pretty print a side for parameters
+prettySide :: HighlightDocument d => Side -> d
+prettySide LHS = text "[left]"
+prettySide RHS = text "[right]"
+
+-- | Pretty print a formal comment
+prettyFormalComment :: HighlightDocument d => String -> String -> d
+prettyFormalComment "" body = multiComment_ [body]
+prettyFormalComment header body = text $ header ++ "{*" ++ body ++ "*}"
+
+
+emptyString :: HighlightDocument d => () -> d
+emptyString _ = text ("")
+
+
+
+-- | Pretty-print a non-empty bunch of intruder rules.
+prettyIntruderVariants :: HighlightDocument d => [IntrRuleAC] -> d
+prettyIntruderVariants vs = vcat . intersperse (text "") $ map prettyIntrRuleAC vs
+
+{-
+-- | Pretty-print the intruder variants section.
+prettyIntrVariantsSection :: HighlightDocument d => [IntrRuleAC] -> d
+prettyIntrVariantsSection rules =
+    prettyFormalComment "section" " Finite Variants of the Intruder Rules " $--$
+    nest 1 (prettyIntruderVariants rules)
+-}
+
+
